+++ app/Http/Controllers/MediaController.php
@@ -18,6 +18,12 @@ class MediaController extends Controller
      */
     public function index()
     {
+        if (!auth()->user()->can('list media mgmt')) {
+            return response()->json([
+                'success' => false,
+                'message' => "You don't have permission to access this module.",
+            ]);
+        }
         $folders = Storage::disk('s3')->directories('production');
         return response()->json(['folders' => $folders]);
     }
@@ -33,6 +39,12 @@ class MediaController extends Controller
      */
     public function show($folder)
     {
+        if (!auth()->user()->can('view media mgmt')) {
+            return response()->json([
+                'success' => false,
+                'message' => "You don't have permission to access this module.",
+            ]);
+        }
         $files = Storage::disk('s3')->files("production/$folder");
         return response()->json(['files' => $files]);
     }
@@ -57,6 +69,12 @@ class MediaController extends Controller
      */
     public function store(Request $request)
     {
+        if (!auth()->user()->can('add media mgmt')) {
+            return response()->json([
+                'success' => false,
+                'message' => "You don't have permission to access this module.",
+            ]);
+        }
         $request->validate([
             'file' => 'required|file',
             'folder' => 'nullable|string'
@@ -81,6 +99,12 @@ class MediaController extends Controller
      */
     public function destroy($file)
     {
+        if (!auth()->user()->can('delete media mgmt')) {
+            return response()->json([
+                'success' => false,
+                'message' => "You don't have permission to access this module.",
+            ]);
+        }
         $filePath = "production/{$file}";
         if (Storage::disk('s3')->exists($filePath)) {
             Storage::disk('s3')->delete($filePath);
